move tf implementation to test, merged master
